Refactor: move NewNote to models dir
