+++ src/components/Sidebar/NotesBar.tsx
@@ -12,7 +12,8 @@ import { FcParallelTasks } from "react-icons/fc";
 import { IoIosSearch } from "react-icons/io";
 import { Link, useParams } from "react-router-dom";
 import { useStopShortcut } from "../../hooks/useStopShortcut";
-import { NewNote, NotesAllCards, cardStore } from "../../store/cardStore";
+import { NewNote } from "../../models/NewNote";
+import { NotesAllCards, cardStore } from "../../store/cardStore";
 import ModalEditor from "../Editor/ModalEditor";
 import { SubsideBar } from "./SubsideBar";
 import SubsideButton from "./SubsideBarButton";
